backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
sqlalchemy==2.0.35
python-dotenv==1.0.1
httpx==0.27.2
python-nmap==0.7.1
scapy==2.6.1
networkx==3.3
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.2
reportlab==4.2.5
anthropic==0.39.0
dnspython==2.7.0
cryptography==43.0.1
websockets==13.1
aiosqlite==0.20.0
